Fix index used for links to repeater pages
Because the algorithm was skipping past items without location information, it was causing the index to get off and the links to be incorrect when going to the individual pages for a repeater.  With this change all the repeaters will be in the list at right but some won't be links if there is no location info.  This should keep the index numbers in sync.